Fix build on Xcode 26.4
Bring the project up to date with Xcode 26 / clang's stricter defaults.

build.sh:
- Carthage workaround xcconfig (MACOSX_DEPLOYMENT_TARGET=10.13,
  GCC_TREAT_WARNINGS_AS_ERRORS=NO) for GCDWebServer/OCMock whose
  declared 10.7/10.6 targets fail at link time on libarclite removal.
- Split bootstrap into checkout + build with --platform Mac, then strip
  iOS/tvOS/watchOS shared schemes so Carthage doesn't waste time
  building platforms a Mac-only CLI never needs.
- Capture raw xcodebuild output via tee for diagnosable failures
  (NO_XCPRETTY=1 disables the pretty filter entirely).

Source fixes for stricter clang under -Werror:
- FBFuture+Sync.m: add explicit (void) on a K&R-empty function decl
  and cast dispatch_time_t to NSTimeInterval at the ExtractResult call
  to silence -Wstrict-prototypes / -Wimplicit-int-float-conversion.
- FBSimulatorHIDEvent.m: explicit (double) and (NSUInteger) casts for
  narrowing and -Wenum-enum-conversion.

+++ FBSimulatorControl/HID/FBSimulatorHIDEvent.m
@@ -212,7 +212,7 @@ static NSString *const KeyY = @"y";
       describeFormat:@"%@ is not a valid event type", typeString]
       fail:error];
   }
-  return [[self alloc] initWithDirection:type x:x.unsignedIntegerValue y:y.unsignedIntegerValue];
+  return [[self alloc] initWithDirection:type x:(double)x.unsignedIntegerValue y:(double)y.unsignedIntegerValue];
 }
 
 - (id)jsonSerializableRepresentation
@@ -355,7 +355,7 @@ static NSString *const ButtonSiri = @"siri";
 
 - (NSUInteger)hash
 {
-  return self.type ^ self.button;
+  return (NSUInteger)self.type ^ (NSUInteger)self.button;
 }
 
 + (NSString *)buttonStringFromButton:(FBSimulatorHIDButton)button
